Moved `core/support/definitions.h` to `core/definitions.h`
