+++ src/lib/shim/shim.h
@@ -6,7 +6,7 @@
 #include <sys/ucontext.h>
 
 #include "lib/shadow-shim-helper-rs/shim_helper.h"
-#include "main/core/support/definitions.h"
+#include "main/core/definitions.h"
 
 // Ensures that the shim is initialized for the current thread. Ok and cheap to
 // call again. Any syscalls made by a process before this is called will not be
