+++ src/main/core/definitions.h
