+++ src/main/host/descriptor/socket.c
@@ -8,7 +8,7 @@
 #include <netinet/in.h>
 #include <sys/un.h>
 
-#include "main/core/support/definitions.h"
+#include "main/core/definitions.h"
 #include "main/core/worker.h"
 #include "main/host/descriptor/compat_socket.h"
 #include "main/host/descriptor/descriptor.h"
