+++ src/main/host/futex.c
@@ -10,7 +10,7 @@
 #include <stdbool.h>
 
 #include "lib/logger/logger.h"
-#include "main/core/support/definitions.h"
+#include "main/core/definitions.h"
 #include "main/core/worker.h"
 #include "main/host/syscall_types.h"
 #include "main/utility/utility.h"
