+++ src/main/host/network/network_interface.c
@@ -10,7 +10,7 @@
 
 #include "lib/logger/logger.h"
 #include "main/bindings/c/bindings.h"
-#include "main/core/support/definitions.h"
+#include "main/core/definitions.h"
 #include "main/core/worker.h"
 #include "main/host/descriptor/compat_socket.h"
 #include "main/host/descriptor/descriptor.h"
