+++ src/main/host/network/network_interface.h
@@ -13,7 +13,7 @@
 typedef struct _NetworkInterface NetworkInterface;
 
 #include "main/bindings/c/bindings.h"
-#include "main/core/support/definitions.h"
+#include "main/core/definitions.h"
 #include "main/host/descriptor/socket.h"
 #include "main/host/protocol.h"
 #include "main/routing/address.h"
