+++ src/main/host/tracker.h
@@ -11,7 +11,7 @@
 #include <netinet/in.h>
 
 #include "lib/logger/log_level.h"
-#include "main/core/support/definitions.h"
+#include "main/core/definitions.h"
 #include "main/host/descriptor/compat_socket.h"
 #include "main/host/protocol.h"
 #include "main/host/tracker_types.h"
