+++ src/main/routing/packet.h
@@ -13,7 +13,7 @@
 #include "main/routing/packet.minimal.h"
 
 #include "main/bindings/c/bindings-opaque.h"
-#include "main/core/support/definitions.h"
+#include "main/core/definitions.h"
 #include "main/host/protocol.h"
 #include "main/host/syscall_types.h"
